Added main.css and index.html
General Description
Implement your task and push to the github repository. Create a simple HTML page with input for input parameters and button to calculate results. After button click javascript code should read input parameters and pass to the function that you will create (based on your task).You should print results. Host your project on github and share a link with a mentor (More details how to host here.). 
Note: who would not be able to solve the task, would need to implement a merge sort algorithm (https://en.wikipedia.org/wiki/Merge_sort).
Task 2 (Pentagonal Number, Combinatorics Challenge).
Have the function CombinatoricsChallenge(num) read num which will be a positive integer and determine how many dots exist in a pentagonal shape around a center dot on the Nth iteration. For example, in the image below you can see that on the first iteration there is only a single dot, on the second iteration there are 6 dots, on the third there are 16 dots, and on the fourth there are 31 dots.